LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/stretch/stretch-block-size-003-expected.txt




















FAIL .test 1 assert_equals:
<div class="test htb" data-expected-height="55"></div>
height expected 55 but got 10
FAIL .test 2 assert_equals:
<div class="test vlr" data-expected-width="41"></div>
width expected 41 but got 10
FAIL .test 3 assert_equals:
<div class="test vrl" data-expected-width="41"></div>
width expected 41 but got 10
FAIL .test 4 assert_equals:
<div class="test htb" data-expected-height="55"></div>
height expected 55 but got 10
FAIL .test 5 assert_equals:
<div class="test vlr" data-expected-width="41"></div>
width expected 41 but got 10
FAIL .test 6 assert_equals:
<div class="test vrl" data-expected-width="41"></div>
width expected 41 but got 10
FAIL .test 7 assert_equals:
<div class="test htb" data-expected-height="41"></div>
height expected 41 but got 10
FAIL .test 8 assert_equals:
<div class="test vlr" data-expected-width="55"></div>
width expected 55 but got 10
FAIL .test 9 assert_equals:
<div class="test vrl" data-expected-width="55"></div>
width expected 55 but got 10
FAIL .test 10 assert_equals:
<div class="test htb" data-expected-height="41"></div>
height expected 41 but got 10
FAIL .test 11 assert_equals:
<div class="test vlr" data-expected-width="55"></div>
width expected 55 but got 10
FAIL .test 12 assert_equals:
<div class="test vrl" data-expected-width="55"></div>
width expected 55 but got 10
FAIL .test 13 assert_equals:
<div class="test htb" data-expected-height="41"></div>
height expected 41 but got 10
FAIL .test 14 assert_equals:
<div class="test vlr" data-expected-width="55"></div>
width expected 55 but got 10
FAIL .test 15 assert_equals:
<div class="test vrl" data-expected-width="55"></div>
width expected 55 but got 10
FAIL .test 16 assert_equals:
<div class="test htb" data-expected-height="41"></div>
height expected 41 but got 10
FAIL .test 17 assert_equals:
<div class="test vlr" data-expected-width="55"></div>
width expected 55 but got 10
FAIL .test 18 assert_equals:
<div class="test vrl" data-expected-width="55"></div>
width expected 55 but got 10
FAIL .test 19 assert_equals:
<div class="test htb" data-expected-height="55" style="box-sizing: border-box;"></div>
height expected 55 but got 10
FAIL .test 20 assert_equals:
<div class="test vlr" data-expected-width="41" style="box-sizing: border-box;"></div>
width expected 41 but got 10
FAIL .test 21 assert_equals:
<div class="test vrl" data-expected-width="41" style="box-sizing: border-box;"></div>
width expected 41 but got 10
FAIL .test 22 assert_equals:
<div class="test htb" data-expected-height="55" style="box-sizing: border-box;"></div>
height expected 55 but got 10
FAIL .test 23 assert_equals:
<div class="test vlr" data-expected-width="41" style="box-sizing: border-box;"></div>
width expected 41 but got 10
FAIL .test 24 assert_equals:
<div class="test vrl" data-expected-width="41" style="box-sizing: border-box;"></div>
width expected 41 but got 10
FAIL .test 25 assert_equals:
<div class="test htb" data-expected-height="41" style="box-sizing: border-box;"></div>
height expected 41 but got 10
FAIL .test 26 assert_equals:
<div class="test vlr" data-expected-width="55" style="box-sizing: border-box;"></div>
width expected 55 but got 10
FAIL .test 27 assert_equals:
<div class="test vrl" data-expected-width="55" style="box-sizing: border-box;"></div>
width expected 55 but got 10
FAIL .test 28 assert_equals:
<div class="test htb" data-expected-height="41" style="box-sizing: border-box;"></div>
height expected 41 but got 10
FAIL .test 29 assert_equals:
<div class="test vlr" data-expected-width="55" style="box-sizing: border-box;"></div>
width expected 55 but got 10
FAIL .test 30 assert_equals:
<div class="test vrl" data-expected-width="55" style="box-sizing: border-box;"></div>
width expected 55 but got 10
FAIL .test 31 assert_equals:
<div class="test htb" data-expected-height="41" style="box-sizing: border-box;"></div>
height expected 41 but got 10
FAIL .test 32 assert_equals:
<div class="test vlr" data-expected-width="55" style="box-sizing: border-box;"></div>
width expected 55 but got 10
FAIL .test 33 assert_equals:
<div class="test vrl" data-expected-width="55" style="box-sizing: border-box;"></div>
width expected 55 but got 10
FAIL .test 34 assert_equals:
<div class="test htb" data-expected-height="41" style="box-sizing: border-box;"></div>
height expected 41 but got 10
FAIL .test 35 assert_equals:
<div class="test vlr" data-expected-width="55" style="box-sizing: border-box;"></div>
width expected 55 but got 10
FAIL .test 36 assert_equals:
<div class="test vrl" data-expected-width="55" style="box-sizing: border-box;"></div>
width expected 55 but got 10
